+++ math/v3/clockface/clockface.go
@@ -6,8 +6,8 @@ import (
 )
 
 type Point struct {
-	X int
-	Y int
+	X float64
+	Y float64
 }
 
 func SecondHand(t time.Time) Point {
